Code improvements: use const and modern type casting where needed

+++ src/widgets/container.cpp
@@ -91,8 +91,8 @@ namespace gcn
         Color faceColor = getBaseColor();
         Color highlightColor, shadowColor;
         int alpha = getBaseColor().a;
-        int width = getWidth() + getBorderSize() * 2 - 1;
-        int height = getHeight() + getBorderSize() * 2 - 1;
+        int width = getWidth() + static_cast<int>(getBorderSize()) * 2 - 1;
+        int height = getHeight() + static_cast<int>(getBorderSize()) * 2 - 1;
         highlightColor = faceColor + 0x303030;
         highlightColor.a = alpha;
         shadowColor = faceColor - 0x303030;
